# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: KS OS External API
  version: 2026-07-01
servers:
  - url: https://api.example.com/api/external/v1
security:
  - bearerApiKey: []
paths:
  /auth/test:
    get:
      operationId: testAuthentication
      responses:
        '200': { description: Credential is valid }
        '401': { $ref: '#/components/responses/Error' }
  /bookings:
    get:
      operationId: listBookings
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { name: cursor, in: query, schema: { type: string, format: date-time } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200':
          description: Tenant-isolated bookings
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Booking' } }
  /services:
    get:
      operationId: listServices
      responses:
        '200': { description: Active tenant services }
components:
  securitySchemes:
    bearerApiKey: { type: http, scheme: bearer, bearerFormat: KS-OS-API-Key }
  schemas:
    Booking:
      type: object
      required: [id, start_time, end_time, status, updated_at]
      properties:
        id: { type: string, format: uuid }
        start_time: { type: string, format: date-time }
        end_time: { type: string, format: date-time }
        status: { type: string }
        updated_at: { type: string, format: date-time }
        service: { type: [string, 'null'] }
        staff: { type: [string, 'null'] }
        location: { type: [string, 'null'] }
  responses:
    Error:
      description: Standard API error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: { type: string }
                  message: { type: string }
                  requestId: { type: string }